Add modified NPM packages for demo purposes
It is necessary to patch isarray packages so that they report on the
version being used when they are called. That way, the users can see
that the versions of dependencies are being correctly resolved.

This packages need to be published to the global NPM registry if they 
are modified so that they can be correctly pulled from the examples.